llm/requirements_rag.txt
flask
transformers
sentence-transformers
faiss-cpu
PyPDF2
numpy
torch
